All admin and user tables are now visible and functional: - User pending reservations - User current loans - User history of loans
- Admin pending reservations
- Admin active loans
-> Admin can now end loans
