+++ src/pages/Employees.jsx
@@ -3,8 +3,8 @@ import { useNavigate } from "react-router";
 import { MdLibraryAdd } from "react-icons/md";
 import AuthContext from "../store/auth-context";
 import "../styles/employees.css";
-import { TextInput } from "./TextInput";
-import { SortedTable } from "./SortedTable";
+import { TextInput } from "../components/TextInput";
+import { SortedTable } from "../components/SortedTable";
 
 function Employees() {
   const auth = useContext(AuthContext);
